Fix Gigya API error 500000 when attempting to save system configuration in a store scope when the default scope does not have a valid Secret key
Mage::getStoreConfig finds configuration in the store scope of the request. But in the admin, this will always be the admin store, including on system configuration pages. Instead, getWebsite->getConfig or getStore->getConfig needs to be used to retrieve current configuration values when on system configuration pages in a website or store scope.