Fix an error when setting user management configuration to Magento + Social Login in a store scope.
Enabling Magento + Social Login causes the social helper to retrieve several configuration values, but it did this using a Mage::getStoreConfig call, which retrieved the configuration in the admin store scope. This caused the API call to account.getSchema to be malformed when the default scope did not have a user key set up and user key was enabled (which is by default), which caused the response to be malformed and missing a callId.