Modify Gigya User Management to be disabled by default
This is a more multitenant-friendly default setting. It should not be presumed that all websites in the application want social login enabled. Also, enabling by default causes non-working social login to be displayed on the login pages as soon as installation of the code happens, before API key configuration can be done.